+++ test/uk/gov/hmrc/gform/core/TemplateValidatorSpec.scala
@@ -772,7 +772,6 @@ class TemplateValidatorSpec extends Spec {
         None,
         None,
         None,
-        None,
         formComponents,
         None,
         None,
