+++ test/uk/gov/hmrc/gform/core/parsers/ValueParserSpec.scala
@@ -516,7 +516,6 @@ class ValueParserSpec extends Spec with TableDrivenPropertyChecks {
       None,
       None,
       None,
-      None,
       List(
         FormComponent(
           FormComponentId("firstName"),
