+++ test/uk/gov/hmrc/gform/formtemplate/FormTemplateSupport.scala
@@ -69,7 +69,6 @@ trait FormTemplateSupport {
         None,
         None,
         None,
-        None,
         formComponents,
         None,
         None,
@@ -237,7 +236,6 @@ trait FormTemplateSupport {
         None,
         None,
         None,
-        None,
         formComponents,
         None,
         None,
@@ -260,7 +258,6 @@ trait FormTemplateSupport {
       Page(
         toSmartString(name),
         pageId,
-        None,
         None,
         None,
         None,
