+++ test/uk/gov/hmrc/gform/sharedmodel/ExampleData.scala
@@ -364,7 +364,7 @@ trait ExampleValidator {
 trait ExampleSection { dependecies: ExampleFieldId with ExampleFieldValue with ExampleValidator =>
   def nonRepeatingPageSection(
     title: String = "About you",
-    validators: Option[Validators] = Some(defaultValidator),
+    validators: Option[Validator] = Some(defaultValidator),
     fields: List[FormComponent] = List(`fieldValue - firstName`, `fieldValue - surname`, `fieldValue - facePhoto`),
     includeIf: Option[IncludeIf] = None
   ) =
@@ -377,7 +377,6 @@ trait ExampleSection { dependecies: ExampleFieldId with ExampleFieldValue with E
         None,
         None,
         includeIf,
-        None,
         validators,
         fields,
         None,
@@ -406,7 +405,6 @@ trait ExampleSection { dependecies: ExampleFieldId with ExampleFieldValue with E
     Section.RepeatingPage(
       Page(
         toSmartString("Repeating section"),
-        None,
         None,
         None,
         None,
