+++ test/uk/gov/hmrc/gform/sharedmodel/formtemplate/generators/SectionGen.scala
@@ -118,8 +118,7 @@ trait SectionGen {
       caption          <- Gen.option(smartStringGen)
       shortName        <- Gen.option(smartStringGen)
       includeIf        <- Gen.option(IncludeIfGen.includeIfGen)
-      validator        <- Gen.option(ValidatorGen.validatorGen)
-      validators       <- Gen.option(ValidatorsGen.validatorsGen)
+      validators       <- Gen.option(ValidatorGen.validatorGen)
       fields           <- PrimitiveGen.oneOrMoreGen(FormComponentGen.formComponentGen())
       continueLabel    <- Gen.option(smartStringGen)
       continueIf       <- Gen.option(ContinueIfGen.continueIfGen)
@@ -136,7 +135,6 @@ trait SectionGen {
       caption,
       shortName,
       includeIf,
-      validator,
       validators,
       fields.toList,
       continueLabel,
